Corrected format of files and folders
