+++ code/classes/house.py
@@ -1,0 +1,16 @@
+class House():
+    """_summary_
+    """
+    def __init__(self, pos, max_output):
+        self.pos = pos
+        self.max_output = max_output
+        self.cables = []
+
+    def lay_cable(self, pos):
+        """_summary_
+
+        Args:
+            pos_x (_type_): _description_
+            pos_y (_type_): _description_
+        """
+        self.cables.append()
