+++ data/district_1/district-1_batteries.csv
@@ -1,0 +1,6 @@
+positie,capaciteit
+"38,12",1507.0
+"43,13",1507.0
+"42,3",1507.0
+"49,23",1507.0
+"3,45",1507.0
