+++ data/district_1/district-1_houses.csv
@@ -1,0 +1,151 @@
+x,y,maxoutput
+34,47,53.97543253
+24,22,66.13020717
+49,44,39.82123377
+19,38,46.4333204
+22,46,62.43058175
+17,11,62.14055349
+19,4,52.43116675
+8,6,30.48448777
+9,37,63.19785578
+16,38,65.00367928
+20,11,40.20336825
+19,12,53.03976583
+33,0,39.13247939
+33,7,39.45690812
+21,37,53.84443114
+37,28,60.40913294
+2,8,27.78076375
+23,48,57.77286486
+28,23,38.72179636
+27,22,29.92408755
+47,6,41.78571625
+38,22,59.91115918
+47,40,71.42208448
+23,1,37.23961036
+18,37,70.21473166
+30,37,61.05026554
+4,33,38.13617191
+16,48,29.83002147
+12,5,74.15296092
+31,41,68.6532457
+50,35,54.25835415
+9,48,41.7880369
+22,10,72.88165705
+26,30,62.0133643
+15,44,53.08761347
+5,15,52.08494864
+21,50,38.71047582
+30,12,66.05341632
+21,29,45.33453242
+37,23,51.94608562
+27,3,57.11814462
+23,10,29.34726889
+13,27,37.28002017
+48,4,58.90934923
+20,14,45.14508246
+11,4,44.75870431
+43,32,46.63929091
+2,50,29.93133051
+30,2,60.7690834
+28,37,55.74063546
+41,36,60.96696147
+4,3,45.9393011
+15,4,70.02856894
+37,47,26.27681146
+8,3,31.03872389
+30,4,29.51140333
+11,10,63.87594515
+4,38,28.04976611
+47,16,31.4973607
+1,18,45.25994244
+30,16,49.93687743
+21,42,65.6434655
+9,50,38.5524821
+0,16,48.15325636
+45,1,59.63813712
+9,44,44.14228862
+44,27,28.73871494
+48,50,35.97663002
+5,13,57.19809379
+46,42,66.07952654
+2,12,75.8473443
+31,18,38.85647001
+45,24,31.97506792
+5,9,38.37311881
+40,39,40.72933179
+3,9,53.27408283
+18,15,57.15314373
+13,28,56.14106998
+50,3,60.5787997
+15,50,47.17876571
+25,5,55.93567156
+25,40,67.70113184
+32,9,32.56142196
+18,28,55.75694271
+43,14,59.68412189
+43,39,30.57764452
+29,22,56.03511409
+21,27,38.69411298
+18,33,64.25056806
+11,38,74.82543524
+16,41,34.78697213
+21,44,46.65642768
+19,40,60.03957103
+37,35,29.26065385
+10,47,70.68631843
+0,34,74.53947279
+10,39,46.25506287
+30,31,33.30314048
+47,50,70.41480105
+39,43,39.37298309
+39,19,64.25603513
+33,11,66.45865072
+24,18,38.3766668
+42,47,26.65336877
+18,32,32.63123071
+5,37,68.63304898
+29,30,62.33504063
+34,50,41.71675352
+44,33,51.84271486
+28,3,29.90737428
+3,7,58.40033018
+7,25,38.50460131
+28,15,37.81128037
+46,38,73.40226925
+38,21,33.56390131
+26,37,33.96614241
+15,46,63.30973309
+39,35,26.81272998
+21,11,47.29043731
+1,24,54.40644245
+20,3,67.12368936
+31,15,39.85726456
+9,23,29.43663895
+42,40,41.46231813
+38,18,69.53106382
+27,13,72.11411933
+34,8,41.23924682
+16,3,33.61818802
+20,34,44.9319234
+38,24,50.53984719
+30,25,63.13649746
+50,49,38.24420236
+24,30,76.1538049
+4,5,43.28587772
+29,7,60.8375359
+36,41,29.1671214
+42,34,43.0013272
+19,3,70.34144168
+24,24,52.07076394
+24,36,63.12900053
+4,0,29.86780761
+19,33,66.97873982
+22,14,42.32972013
+37,6,30.41692196
+33,24,74.26615041
+39,14,49.16852893
+21,48,71.35472201
+22,17,66.78438663
+26,27,56.65609954
+10,27,26.19979919
