+++ data/district_2/district-2_batteries.csv
@@ -1,0 +1,6 @@
+positie,capaciteit
+"19,20",1508.25
+"1,36",1508.25
+"34,49",1508.25
+"41,21",1508.25
+"26,22",1508.25
