+++ data/district_2/district-2_houses.csv
@@ -1,0 +1,151 @@
+x,y,maxoutput
+44,6,48.76400191
+13,35,44.09168174
+27,6,38.4723184
+14,21,36.40331775
+50,40,44.36102572
+15,40,55.71262043
+36,36,45.9622578
+3,4,36.90684143
+26,48,49.9985406
+44,18,63.12038188
+32,41,63.7118043
+33,5,56.91378259
+31,19,40.60892953
+4,1,35.82814756
+48,26,36.88169378
+10,4,64.52218421
+14,40,39.38884807
+17,9,49.84931475
+47,30,61.0137363
+16,47,62.28650955
+32,4,36.98224478
+8,19,64.06758851
+15,19,41.79763764
+32,8,56.03098521
+45,12,64.50983467
+50,29,53.17296019
+4,40,36.37710418
+40,32,55.38231616
+48,41,61.3949773
+36,32,52.84858044
+25,5,37.73044496
+41,2,47.48157843
+50,18,36.13312321
+21,37,47.33240149
+45,50,57.24475908
+40,29,63.05172666
+5,29,40.77396592
+5,3,54.13000914
+33,28,57.01992312
+8,42,49.57599547
+9,26,48.83351279
+31,15,64.4517609
+23,10,39.91374306
+14,48,64.79397046
+49,7,58.44840778
+43,40,52.20518124
+37,18,51.14671974
+0,38,41.12708974
+29,28,48.50865344
+28,5,59.78619507
+20,2,39.31626881
+47,27,58.76139914
+25,16,63.73606908
+1,1,52.204709
+27,5,36.7394839
+26,18,39.8339645
+47,19,40.4929237
+6,9,61.5437326
+17,15,52.23919855
+44,29,57.03641574
+12,10,61.57329114
+19,28,54.48982494
+11,34,38.62374353
+49,39,43.1260208
+27,18,44.4066236
+24,19,64.44382811
+12,38,38.40462319
+18,6,64.1795134
+18,44,40.87885768
+29,1,63.87700034
+36,50,47.14949021
+3,9,58.3986013
+50,45,64.49005705
+26,31,62.88141752
+47,35,43.94928406
+29,31,50.35807729
+12,50,56.40032746
+15,45,60.93412813
+5,19,58.6298162
+1,43,47.39904077
+15,4,61.34791216
+17,21,59.05896036
+28,44,59.68436389
+21,29,49.50781665
+1,29,39.81936486
+37,13,41.68226663
+42,28,40.61631674
+41,12,55.90292767
+4,44,42.61411328
+28,46,49.44122897
+50,12,49.04859791
+3,18,57.36334288
+37,12,38.33821095
+9,4,54.0265882
+1,12,57.39706274
+17,35,43.25342339
+8,33,49.04355789
+17,44,63.56837786
+15,29,58.39362925
+12,11,38.79808279
+33,32,35.44113109
+10,25,46.87020086
+3,35,64.6667857
+2,19,52.89669021
+36,12,36.64489815
+18,18,38.48596722
+14,11,47.63820126
+48,3,63.8665196
+2,18,62.12319013
+30,35,50.13812647
+1,9,45.04775519
+16,50,39.51931013
+14,4,64.74853094
+11,9,45.54974087
+37,3,38.76288176
+24,29,54.907644
+13,33,57.5929688
+1,5,54.51204967
+31,13,45.8934274
+21,48,41.48920914
+0,26,43.44180597
+4,38,39.14559958
+24,26,38.69040639
+46,27,38.17991123
+50,31,50.56612579
+20,32,54.8377261
+11,15,36.67499817
+23,34,41.88875758
+7,29,52.8837846
+46,34,44.36408371
+6,42,40.7044132
+25,37,48.28655559
+4,16,47.83949256
+10,27,51.50674585
+33,47,59.37576043
+8,50,50.04026027
+8,8,49.40039856
+41,34,55.9303441
+12,41,41.16116018
+37,37,50.75214489
+16,29,60.91065902
+43,11,40.55434782
+5,35,44.58706876
+45,30,58.83768219
+32,38,56.32473588
+20,37,54.95406591
+38,47,35.5552718
+3,25,38.40802103
+34,1,37.82715244
+40,31,57.12008594
