+++ data/district_3/district-3_batteries.csv
@@ -1,0 +1,6 @@
+positie,capaciteit
+"18,34",1506.75
+"32,11",1506.75
+"41,1",1506.75
+"3,35",1506.75
+"39,41",1506.75
