+++ data/district_3/district-3_houses.csv
@@ -1,0 +1,151 @@
+x,y,maxoutput
+21,18,51.94411629
+47,50,47.82087397
+25,31,51.81241388
+27,38,53.11777253
+30,13,54.14676516
+48,4,45.24272018
+32,18,46.74039375
+49,15,54.75398768
+3,9,47.72196575
+46,40,51.86891316
+3,25,46.8033198
+1,32,54.99730646
+32,3,47.04037014
+32,49,50.25432556
+30,14,45.54596912
+39,34,49.93506537
+3,24,49.07862835
+43,45,54.18469088
+1,28,46.63089313
+15,0,48.43740503
+30,32,52.92581641
+18,10,53.98009231
+21,46,52.91142802
+21,33,46.75613739
+39,42,47.34915773
+17,21,46.26812357
+12,21,53.63368373
+17,34,52.96938435
+46,39,46.60969509
+33,35,45.84094791
+43,41,51.95373609
+15,36,54.30849405
+18,24,54.2211502
+16,37,52.58547404
+47,35,46.9524009
+42,4,53.46156177
+3,33,49.79974209
+41,24,51.75273225
+4,48,48.27297674
+1,45,54.91354462
+27,10,50.04190502
+28,48,54.22966024
+11,23,48.85814641
+1,48,54.54461427
+43,29,47.94582721
+38,3,49.86535747
+31,24,46.38651333
+25,29,53.35585729
+16,19,54.16981224
+45,31,48.01065715
+27,11,50.55356632
+46,48,47.37788381
+17,41,48.60918505
+39,46,48.14242474
+14,6,49.72629004
+8,12,46.01633925
+0,19,51.03867367
+13,47,48.4395487
+39,35,52.0929123
+25,13,45.5472391
+8,13,46.46747877
+33,31,48.25043584
+28,28,45.40186719
+39,25,49.93311335
+12,26,49.65236115
+6,41,46.05322103
+16,1,48.70842308
+20,11,48.54997218
+7,32,51.72717898
+2,5,51.62646163
+41,36,52.78210048
+10,36,49.13738507
+4,2,52.76961749
+20,16,48.0219944
+31,8,45.78474275
+24,50,52.98566399
+25,40,49.79777396
+48,28,47.99198559
+50,0,53.71601636
+30,7,45.2342213
+21,21,53.15980812
+20,10,50.30051709
+6,17,46.01918212
+30,45,50.8435769
+50,15,53.95360569
+16,16,53.32105133
+14,22,50.63918945
+11,37,52.2305101
+50,21,51.82639703
+9,11,52.8088529
+15,10,54.53006205
+6,22,46.51627586
+41,44,45.29702228
+26,41,47.09131203
+15,15,45.31733653
+24,32,50.16548949
+41,12,52.38040091
+29,25,47.58049297
+39,27,52.05437525
+50,35,54.12515118
+0,4,45.16959822
+45,2,53.48855448
+29,7,50.91041056
+10,24,53.76285577
+46,4,47.6825473
+17,12,45.61104522
+33,3,53.3512745
+36,14,51.58057128
+8,44,50.30256407
+48,47,53.26904882
+22,34,46.65187897
+31,37,51.54326497
+19,47,50.38289294
+20,28,53.12331423
+20,5,48.31035497
+19,22,51.79628813
+6,43,53.95936785
+2,50,49.52760828
+8,20,46.13662116
+40,8,45.17557463
+34,47,51.57104914
+0,46,50.20207187
+37,21,54.7990422
+14,15,53.34058355
+44,40,46.43816048
+0,24,48.72125695
+3,23,53.08724525
+17,37,53.63429606
+42,45,49.54607954
+9,40,54.56844455
+37,5,47.92097158
+1,30,48.43719875
+10,3,50.74878243
+36,44,47.56339729
+6,0,47.77817151
+30,5,45.25109877
+3,16,53.54773542
+43,30,47.97674042
+2,35,52.57995149
+30,29,47.67000862
+29,16,48.94518723
+36,8,51.83878558
+25,39,47.71839737
+18,39,48.74438962
+4,35,48.22881096
+6,2,48.23526882
+14,18,51.07946117
+35,24,48.82801588
+20,41,46.27510427
+29,11,47.73944395
